# dist/api.yml
openapi: 3.0.3
info:
  version: '1.0.0'
  title: 'Web Development API'
  description: 'API for assignment 3.1'
...
paths:
...
  /users/favoriteMatches:
    get:
      tags: 
        - users
      summary: Returns all user's favorite matches.
      description: 'returns array of all favorite matches ids'
      responses:
        '200':
          description: 'successful fetched all favourite games ids.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/favourite_matches'
        '404':
          description: 'data not found.'
    post:
      tags: 
        - users
      summary: Add match to favorites.
      description: 'add match to user favourites list.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/favourite_add_match'
      responses:
        '201':
          description: 'succesfully added match to favourites.'
        '404':
          description: 'data not found.'
  /users/favoriteMatches/{match_id}:
    parameters:
        - name: match_id
          required: true
          in: path
          description: 'match id'
          schema:
            type: integer
            example: 8
    delete:
      tags: 
        - users
      summary: delete match from favorites.
      description: delete match from user's favorites given its id
      responses:
        '200':
          description: 'successful deleted match id from favorites.'
        '404':
          description: 'data not found.'
  /users/favoritePlayers:
      get:
        tags: 
          - users
        summary: Returns all user's favorite players.
        description: 'returns array of all favorite players ids'
        responses:
          '200':
            description: 'successful fetched all favourite players ids.'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/favourite_players'
          '404':
            description: 'data not found.'
      post:
        tags: 
          - users
        summary: Add player to favorites.
        description: 'add player to user favourites list.'
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/favourite_add_player'
        responses:
          '201':
            description: 'succesfully added player to favourites.'
          '404':
            description: 'data not found.'
  /users/favoritePlayers/{player_id}:
    delete:
      tags: 
        - users
      summary: delete player from favorites.
      description: delete player from user's favorites given its id
      parameters:
        - name: player_id
          required: true
          in: path
          description: 'player id'
          schema:
            type: integer
            example: 26722
      responses:
        '200':
          description: 'successful deleted player id from favorites.'
        '404':
          description: 'data not found.'
  /users/favoriteTeams:
        get:
          tags: 
            - users
          summary: Returns all user's favorite teams.
          description: 'returns array of all favorite teams ids'
          responses:
            '200':
              description: 'successful fetched all favourite teams ids.'
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/favourite_teams'
            '404':
              description: 'data not found.'
        post:
          tags: 
            - users
          summary: Add team to favorites.
          description: 'add team to user favourites list.'
          requestBody:
            required: true
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/favourite_add_team'
          responses:
            '201':
              description: 'succesfully added team to favourites.'
            '404':
              description: 'data not found.'
  /users/favoriteTeams/{team_id}:
      delete:
        tags: 
          - users
        summary: delete team from favorites.
        description: delete team from user's favorites given its id
        parameters:
          - name: team_id
            required: true
            in: path
            description: 'team id'
            schema:
              type: integer
              example: 85
        responses:
          '200':
            description: 'successful deleted team id from favorites.'
          '404':
            description: 'data not found.'
...
components:
  schemas:
...
    favourite_matches:
      type: array
      items:
        properties:
          match_id:
                type: integer
                description: match id
                example: 8
    favourite_add_match:
      type: object
      properties:
          match_id:
                type: integer
                description: match id
                example: 8
    favourite_players:
      type: array
      items:
        properties:
          player_id:
                type: integer
                description: player id
                example: 26722
    favourite_add_player:
      type: object
      properties:
          player_id:
                type: integer
                description: player id
                example: 26722
    favourite_teams:
      type: array
      items:
        properties:
          team_id:
                type: integer
                description: team id
                example: 85
    favourite_add_team:
      type: object
      properties:
          team_id:
                type: integer
                description: team id
                example: 85
...